Add onionmessenger.com to the blacklist
They do not publish contact information on XMPP or on the web (only known website is a direct converse installation).

I have tried to contact them through their ISP for pedo content almost two months ago and never got an answer back. I believe it is safe to assume that an inquiry regarding spammers (which are present on this service and I saw not later than today) will be met with the same silence.